requirements.txt
asgiref==3.8.1
certifi==2025.4.26
distlib==0.3.9
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
filelock==3.18.0
Markdown==3.7
packaging==25.0
pipenv==2025.0.2
platformdirs==4.3.7
PyJWT==2.9.0
setuptools==80.2.0
sqlparse==0.5.3
virtualenv==20.30.0
